# PERF: Improve performance of cut with IntervalIndex bins